layout: move import components to features/imports/
Now that the import-related components are connected to the Redux store,
they're considered part of a single "feature" and should no longer be in
app/.  Move the import page and its related components into
features/imports/.
